chore: no need for a new hook
